Refactors user modification endpoint.
Updates the user modification endpoint to receive the user ID as a path variable and applies validation to the request body.
This change improves the security and robustness of the user modification process by ensuring that only valid data is used to update user information.
It also throws exception when a user tries to modify other user data by id.
Removes unused mapper method.
Borrows recetaService and versionRecetaService in usuarioService to implement delete user cascade.

+++ src/main/java/com/unlu/alimtrack/services/UsuarioService.java
@@ -20,16 +20,20 @@ public class UsuarioService {
     private final UsuarioRepository usuarioRepository;
     private final UsuarioModelMapper usuarioMapper;
     private final PasswordEncoder passwordEncoder;
+    private final RecetaService recetaService;
+    private final VersionRecetaService versionRecetaModelService;
 
-    public UsuarioService(UsuarioRepository usuarioRepository, UsuarioModelMapper usuarioMapper, PasswordEncoder passwordEncoder) {
+    public UsuarioService(UsuarioRepository usuarioRepository, UsuarioModelMapper usuarioMapper, PasswordEncoder passwordEncoder, RecetaModelRepository recetaModelRepository, RecetaService recetaModelService, VersionRecetaService versionRecetaModelService) {
         this.usuarioRepository = usuarioRepository;
         this.usuarioMapper = usuarioMapper;
         this.passwordEncoder = passwordEncoder;
+        this.recetaService = recetaModelService;
+        this.versionRecetaModelService = versionRecetaModelService;
     }
 
     public List<UsuarioResponseDTO> getAllUsuarios() {
         List<UsuarioModel> usuarios = usuarioRepository.findAll();
-        if  (usuarios.isEmpty()) {
+        if (usuarios.isEmpty()) {
             throw new RecursoNoEncontradoException("No hay usuarios guardados");
         }
         return usuarios.stream().map(
@@ -39,7 +43,7 @@ public class UsuarioService {
     public UsuarioResponseDTO saveUsuario(UsuarioCreateDTO usuario) {
         UsuarioModel usuarioModel = usuarioMapper.usuarioCreateDTOToModel(usuario);
         // verifica si ya existe un usuario con ese email
-        if (usuarioRepository.existsByEmail(usuarioModel.getEmail())){
+        if (usuarioRepository.existsByEmail(usuarioModel.getEmail())) {
             throw new RecursoYaExisteException("El email ya ha sido usado por un usuario existente");
         }
         // crea el usuario y devuelve un response
@@ -49,24 +53,36 @@ public class UsuarioService {
         return usuarioMapper.usuarioToUsuarioResponseDTO(usuarioModel);
     }
 
-    public UsuarioResponseDTO getUsuarioDtoById(Long id) {
+    public UsuarioResponseDTO getUsuarioResponseDTOById(Long id) {
         UsuarioModel usuarioModel = usuarioRepository.findById(id).orElse(null);
         if (usuarioModel == null) {
-            throw new RecursoNoEncontradoException("Usuario no encontrado");
+            throw new RecursoNoEncontradoException("Usuario no encontrado con id " + id);
         }
         return usuarioMapper.usuarioModelToUsuarioResponseDTO(usuarioModel);
     }
 
-    public UsuarioResponseDTO getUsuarioModelById(Long id) {
-        UsuarioModel usuarioModel =  usuarioRepository.findById(id).orElse(null);
-        return usuarioMapper.usuarioModelToUsuarioResponseDTO(usuarioModel);
+    public UsuarioModel getUsuarioModelById(Long id) {
+        UsuarioModel usuarioModel = usuarioRepository.findById(id).orElse(null);
+        if (usuarioModel == null) {
+            throw new RecursoNoEncontradoException("Usuario no encontrado con id " + id);
+        }
+        return usuarioModel;
     }
 
-    public void modificarUsuario(UsuarioModifyDTO modificacion) {
-        if (validarModificacionUsuario(modificacion)){
+    public void modificarUsuario(Long id, UsuarioModifyDTO modificacion) {
+        if (!validarModificacionUsuario(modificacion)) {
             throw new ModificacionInvalidaException("No se puede realizar la modificacion solicitada");
-        };
-        usuarioRepository.save(usuarioMapper.usuarioModifyDTOToModel(modificacion));
+        }
+        UsuarioModel usuario = usuarioRepository.findById(id).orElse(null);
+        if (usuario == null) {
+            throw new RecursoNoEncontradoException("Usuario no encontrado con id " + id);
+        }
+
+        usuario.setNombre(modificacion.nombre());
+        String passwordEncriptada = passwordEncoder.encode(modificacion.contraseña());
+        usuario.setContraseña(passwordEncriptada);
+
+        usuarioRepository.save(usuario);
     }
 
     private boolean validarModificacionUsuario(UsuarioModifyDTO modificacion) {
@@ -74,6 +90,13 @@ public class UsuarioService {
     }
 
     public void borrarUsuario(Long id) {
+        if (recetaService.getRecetaModelById(id) == null) {
+            throw new IllegalStateException("No se puede borrar el usuario, tiene recetas asociadas.");
+        }
+        if (versionRecetaModelService.getVersionById(id) == null ) {
+            throw new IllegalStateException("No se puede borrar el usuario, tiene recetas asociadas.");
+        }
+
         usuarioRepository.deleteById(id);
     }
